Refactor gestaltd bootstrap and server file layout
